+++ proto/merge_test.go
@@ -1,0 +1,14 @@
+package proto_test
+
+import (
+	"testing"
+
+	"github.com/cosmos/gogoproto/proto"
+)
+
+func TestMergedRegistry(t *testing.T) {
+	_, err := proto.MergedRegistry()
+	if err != nil {
+		t.Error(err)
+	}
+}
